Separate fingerprint authentication dialog code
